+++ qom/ui/widgets/SystemWidget.py
@@ -194,16 +194,16 @@ class SystemWidget(BaseWidget):
         # selected function name
         func_name = self.cmbx_func.currentText()
 
-        # update system parameters
-        self.system.params = params
+        # initialize system
+        system = self.system(params=params, cb_update=self.parent.update)
 
         # extract parameters
-        _, c = self.system.get_ivc()
-        _len_D = 4 * self.system.num_modes**2
+        _, c = system.get_ivc()
+        _len_D = 4 * system.num_modes**2
         params = c[_len_D:] if len(c) > _len_D else c
 
         # get value
-        value = getattr(self.system, 'get_' + func_name)(params)
+        value = getattr(system, 'get_' + func_name)(params)
 
         return value
 
@@ -221,7 +221,7 @@ class SystemWidget(BaseWidget):
 
         # iterate through available systems
         for system in self.systems:
-            codes.append(system({}).code)
+            codes.append(system(params={}, cb_update=None).code)
 
         return codes
 
@@ -257,16 +257,18 @@ class SystemWidget(BaseWidget):
         row_height = 32
         pre_count = 7
 
-        # update parameter
-        self.pos = pos
-        self.system = self.systems[pos](params={}, cb_update=self.parent.update)
+        # update system
+        self.system = self.systems[pos]
+
+        # initalize system
+        system = self.system(params={}, cb_update=None)
 
         # search for available functions
-        func_names = [func[4:] for func in dir(self.system) if callable(getattr(self.system, func)) and func[:4] == 'get_']
+        func_names = [func[4:] for func in dir(system) if callable(getattr(system, func)) and func[:4] == 'get_']
         # filter functions with system parameters
         cmbx_items = list()
         for func_name in func_names:
-            func = getattr(self.system, 'get_' + func_name)
+            func = getattr(system, 'get_' + func_name)
             func_args = inspect.getfullargspec(func).args[1:]
             if len(func_args) == 1 and func_args[0] == 'params':
                 cmbx_items.append(func_name)
@@ -283,7 +285,7 @@ class SystemWidget(BaseWidget):
         self.param_widgets = list()
 
         # add widgets
-        params = self.system.params
+        params = system.params
         widget_col = 0
         for param in params:
             # new widget
@@ -297,7 +299,7 @@ class SystemWidget(BaseWidget):
             widget_col += 1
 
         # update UI elements
-        self.lbl_name.setText(self.system.name)
+        self.lbl_name.setText(system.name)
         self.lbl_func.setVisible(True)
         self.cmbx_func.clear()
         self.cmbx_func.addItems(cmbx_items)