+++ qom/utils/looper.py
@@ -15,20 +15,20 @@ from ..loopers import XLooper, XYLooper, XYZLooper
 default_looper_func_names = {
     'averaged_eigenvalues': 'aes',
     'lyapunov_exponents': 'les', 
-    'measure_averages': 'mes',
+    'measure_average': 'mav',
     'measure_dynamics': 'mdy',
     'mean_optical_occupancy': 'moo',
     'pearson_correlation_coefficient': 'pcc'
 }
 
-def get_looper_func(system, solver_params: dict, func_code: str):
+def get_looper_func(SystemClass, solver_params: dict, func_code: str):
     """Function to obtain to the looper function.
 
     Requires already defined callables ``func_ode``, ``get_mode_rates``, ``get_ivc``, ``get_A`` and ``get_oss_args`` inside the system class.
     
     Parameters
     ----------
-    system : :class:`qom.systems.*`
+    SystemClass : :class:`qom.systems.*`
         Class containing the system.
     solver_params : dict
         All parameters of the solver.
@@ -48,8 +48,8 @@ def get_looper_func(system, solver_params: dict, func_code: str):
 
     # function to calculate averaged eigenvalues of the drift matrix
     def func_averaged_eigenvalues(system_params, val, logger, results):
-        # update system
-        system.params = system_params
+        # initialize system
+        system = SystemClass(system_params)
         # get averaged eigenvalues
         aes = system.get_averaged_eigenvalues(solver_params=solver_params)
         # update results
@@ -57,8 +57,8 @@ def get_looper_func(system, solver_params: dict, func_code: str):
 
     # function to calculate measure average
     def func_measure_average(system_params, val, logger, results):
-        # update system
-        system.params = system_params
+        # initialize system
+        system = SystemClass(system_params)
         # get measure average
         mav = system.get_measure_average(solver_params=solver_params)
         # update results
@@ -66,8 +66,8 @@ def get_looper_func(system, solver_params: dict, func_code: str):
 
     # function to calculate measure dynamics
     def func_measure_dynamics(system_params, val, logger, results):
-        # update system
-        system.params = system_params
+        # initialize system
+        system = SystemClass(system_params)
         # get measure dynamics
         mdy, _ = system.get_measure_dynamics(solver_params=solver_params)
         # update results
@@ -75,8 +75,8 @@ def get_looper_func(system, solver_params: dict, func_code: str):
 
     # function to obtain the maximum Lyapunov exponent
     def func_lyapunov_exponents(system_params, val, logger, results):
-        # update system
-        system.params = system_params
+        # initialize system
+        system = SystemClass(system_params)
         # get Lyapunov exponents
         les = system.get_lyapunov_exponents(solver_params=solver_params)
         # update results
@@ -84,8 +84,8 @@ def get_looper_func(system, solver_params: dict, func_code: str):
 
     # function to calculate mean optical occupancies
     def func_mean_optical_occupanies(system_params, val, logger, results):
-        # update system
-        system.params = system_params
+        # initialize system
+        system = SystemClass(system_params)
         # get mean optical occupancies
         moo, _ = system.get_mean_optical_occupancies()
         # update results
@@ -93,8 +93,8 @@ def get_looper_func(system, solver_params: dict, func_code: str):
 
     # function to calculate Pearson correlation coefficient
     def func_pearson_correlation_coefficient(system_params, val, logger, results):
-        # update system
-        system.params = system_params
+        # initialize system
+        system = SystemClass(system_params)
         # get Pearson correlation coefficient
         pcc = system.get_pearson_correlation_coefficient(solver_params=solver_params)
         # update results
@@ -104,8 +104,8 @@ def get_looper_func(system, solver_params: dict, func_code: str):
     def get_func_params(func_name):
         # functio to calculate
         def func_params(system_params, val, logger, results):
-            # update system
-            system.params = system_params
+            # initialize system
+            system = SystemClass(system_params)
             # extract parameters
             _, c = system.get_ivc()
             _len_D = 4 * system.num_modes**2
@@ -116,6 +116,9 @@ def get_looper_func(system, solver_params: dict, func_code: str):
             results.append((val, value))
         
         return func_params
+
+    if func_code in default_looper_func_names:
+        func_code = default_looper_func_names[func_code]
 
     func = {
         'aes': func_averaged_eigenvalues,
@@ -128,14 +131,14 @@ def get_looper_func(system, solver_params: dict, func_code: str):
 
     return func
 
-def wrap_looper(system, params: dict, func, looper, file_path: str=None, plot: bool=False, hold: bool=True, width: float=5.0, height: float=5.0):
+def wrap_looper(SystemClass, params: dict, func, looper, file_path: str=None, plot: bool=False, hold: bool=True, width: float=5.0, height: float=5.0):
     """Function to wrap loopers.
 
     Requires already defined callables ``func_ode``, ``get_mode_rates``, ``get_ivc``, ``get_A`` and ``get_oss_args`` inside the system class.
     
     Parameters
     ----------
-    system : :class:`qom.systems.*`
+    SystemClass : :class:`qom.systems.*`
         Class containing the system.
     params : dict
         All parameters as defined in :class:`qom.loopers.BaseLooper`.
@@ -180,13 +183,9 @@ def wrap_looper(system, params: dict, func, looper, file_path: str=None, plot: b
     # initialize logger
     init_log()
 
-    # update system parameters
-    system.params.update(params['system'])
-    params['system'] = system.params
-
     # select function
     if type(func) is str:
-        func = get_looper_func(system, params, func)
+        func = get_looper_func(SystemClass=SystemClass, solver_params=params['solver'], func_code=func)
 
     # select looper
     if type(looper) is str: